% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{bs_put}
\alias{bs_put}
\title{Black-Scholes price of a European put}
\usage{
bs_put(S0, K, r, sigma, maturity)
}
\arguments{
\item{S0}{numeric(1), spot price, > 0}

\item{K}{numeric(1), strike price, > 0}

\item{r}{numeric(1), continuously compounded riskless rate}

\item{sigma}{numeric(1), volatility, > 0}

\item{maturity}{numeric(1), time to expiry in years, > 0}
}
\value{
numeric(1), the arbitrage-free put price
}
\description{
Black-Scholes price of a European put
}
